requirements.txt
darts==0.32.0
dask==2025.1.0
gymnasium==1.0.0
ipython==8.12.3
joblib==1.4.2
matplotlib==3.8.3
numpy==2.2.2
optuna==4.2.0
pandas==2.2.3
pandas_market_calendars==4.6.1
psutil==6.1.1
pytorch_lightning==2.5.0.post0
scikit_learn==1.6.1
seaborn==0.13.2
stock_indicators==1.3.3
streamlit==1.36.0
torch==2.5.1+cu118
tqdm==4.67.1
yfinance==0.2.52
google-generativeai==0.3.2
